prepare for build with win32

+++ src/c/ngx_http_clojure_mem.c
@@ -802,7 +802,7 @@ int ngx_http_clojure_eval(int cid, void *r) {
 	int rc;
 	log_debug1(ngx_http_clojure_global_cycle->log, "ngx clojure eval request: %ul", (uintptr_t)r);
 	log_debug2(ngx_http_clojure_global_cycle->log, "ngx clojure eval request to jlong: %" PRIu64 ", size: %d", (jlong)(uintptr_t)r, 8);
-	rc = (*env)->CallStaticIntMethod(env, nc_rt_class,  nc_rt_eval_mid, (jint)cid, (uintptr_t)r);
+	rc = (*env)->CallStaticIntMethod(env, nc_rt_class,  nc_rt_eval_mid, (jint)cid, (jlong)(uintptr_t)r);
 	exception_handle(1, env, return 500);
 	return rc;
 }
